requirements.txt
langchain
langchain_community
chainlit==1.0.506
langchain-groq
groq
langchain-openai
python-dotenv
huggingface_hub
chromadb
openai
pymupdf4llm
langchain_nvidia_ai_endpoints